requirements.txt
python-jobspy>=1.0.5
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
pandas>=2.0.0 